Add files via upload
ZIP file with all of the jupyter notebooks containing the ML models and CSV used to train them. The machine learning models found in this file include: a K-nearest neighbours model, Support Vectors Classifier model and two Random Forest models (one of them being a replica of the article's model).